libs/cast-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cast-core"
version = "0.1.0"
description = "Core library for Cast Sync - parsing, normalization, digests"
requires-python = ">=3.10"
dependencies = [
    "ruamel.yaml>=0.18.0",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0"
]